% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCStan-class.R
\docType{class}
\name{BASiCStan-class}
\alias{BASiCStan-class}
\title{Thin wrapper for BASiCS models in \linkS4class{stanfit} form before conversion.}
\description{
Thin wrapper for BASiCS models in \linkS4class{stanfit} form before conversion.
}
\section{Slots}{

\describe{
\item{\code{FeatureNames,ObservationNames}}{The feature (gene) and observation (cell) names of the data used to fit the model.}
}}

