% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{check_cond}
\alias{check_cond}
\title{check whether the updates of the optimization path should be used in the
inverse Hessian estimation or not}
\usage{
check_cond(Yk, Sk)
}
\arguments{
\item{Yk}{update in parameters}

\item{Skt}{update in gradient}
}
\value{
A logical value
}
\description{
check whether the updates of the optimization path should be used in the
inverse Hessian estimation or not
}
