% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCStan.R
\name{.stan2basics}
\alias{.stan2basics}
\title{Convert stan fits to BASiCS_Chain objects.}
\usage{
.stan2basics(x, gene_names = NULL, cell_names = NULL, size_factors = NULL)
}
\arguments{
\item{x}{A stan object}

\item{gene_names, cell_names}{Gene and cell names. The reason this argument
exists is that by default, stan fit parameters are not named.
NOTE: this must be the same order as the
data supplied to BASiCS_stan.}

\item{size_factors}{Cell-specific scaling normalisation factors, to be
stored as part of the chain object when \code{WithSpikes=FALSE}.}
}
\value{
A BASiCS_Chain object.
}
\description{
Convert stan fits to BASiCS_Chain objects.
}
