% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{est_DIV}
\alias{est_DIV}
\title{estimate divergence based on Monte Carlo samples given the output of
subroutine Form_N_apx}
\usage{
est_DIV(sample_pkg, N_sam, fn, label = "ELBO")
}
\arguments{
\item{sample_pkg}{The output of subroutine Form_N_apx}

\item{N_sam}{number of samples used in estimating divergence}

\item{fn}{negative of log density}

\item{label}{type of divergence, (default = "ELBO")}
}
\value{
A named list with elements:
\code{DIV}
\code{repeat_draws}
\code{fn_draws}
\code{lp_approx_draws}
\code{fn_calls_DIV}
}
\description{
estimate divergence based on Monte Carlo samples given the output of
subroutine Form_N_apx
}
