% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{Form_init_Diag}
\alias{Form_init_Diag}
\title{Form the initial diagonal inverse Hessian in the L-BFGS update}
\usage{
Form_init_Diag(E0, Yk, Sk)
}
\arguments{
\item{E0}{initial diagonal inverse Hessian before updated}

\item{Yk}{update in parameters}

\item{Skt}{update in gradient}
}
\value{
The diagonal inverse Hessian.
}
\description{
Form the initial diagonal inverse Hessian in the L-BFGS update
}
