% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{Form_N_apx_taylor}
\alias{Form_N_apx_taylor}
\title{Returns sampling metrics of the approximating Gaussian given the history
of optimization trajectory}
\usage{
Form_N_apx_taylor(x_l, g_l, Ykt_h, Skt_h, E, lmm)
}
\arguments{
\item{x_l}{The point up to which the optimization path is used for approximation}

\item{g_l}{The gradient at x_l}

\item{Ykt_h}{history of updates along optimization trajectory}

\item{Skt_h}{history of updates of gradients along optimization trajectory}

\item{E}{initial diagonal inverse Hessian}

\item{lmm}{The size of history}
}
\value{
A list with named elements:
\code{label}
\code{theta_D}
\code{Qk}
\code{Rktilde}
\code{logdetcholHk}
\code{Mkbar}
\code{Wkbart}
\code{x_center}
}
\description{
Returns sampling metrics of the approximating Gaussian given the history
of optimization trajectory
}
