% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{Imp_Resam_Each}
\alias{Imp_Resam_Each}
\title{PSIS for one-path Pathfinder}
\usage{
Imp_Resam_Each(param_path, seed)
}
\arguments{
\item{param_path}{output of function opt_path_stan_parallel()}

\item{seed}{random seed for importance resampling}
}
\description{
PSIS for one-path Pathfinder
}
