% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{Imp_Resam_WOR}
\alias{Imp_Resam_WOR}
\title{PSIS-IR without replacement for multi-path Pathfinder
Return n_inits distinct samples}
\usage{
Imp_Resam_WOR(param_path, n_inits, seed = 123)
}
\arguments{
\item{param_path}{output of function opt_path_stan_parallel()}

\item{n_inits}{number of distinct samples}

\item{seed}{random seed for importance resampling}
}
\description{
PSIS-IR without replacement for multi-path Pathfinder
Return n_inits distinct samples
}
