% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{Imp_Resam_WR}
\alias{Imp_Resam_WR}
\title{PSIS-IR with replacement for multi-path Pathfinder
Return n_sam samples}
\usage{
Imp_Resam_WR(param_path, n_sam, seed = 123)
}
\arguments{
\item{param_path}{output of function opt_path_stan_parallel()}

\item{n_sam}{number of distinct samples}

\item{seed}{random seed for importance resampling}
}
\description{
PSIS-IR with replacement for multi-path Pathfinder
Return n_sam samples
}
