% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{opt_path_stan}
\alias{opt_path_stan}
\title{Run one-path Pathfinder for specified Stan model and data for
the specified number of iterations using the specified bound on
uniform initialization on the unconstrained scale.  See opt_path()
for a description of output format and algorithm.}
\usage{
opt_path_stan(
  model,
  data,
  init_bound = 2,
  N1 = 1000,
  N_sam_DIV = 5,
  N_sam = 100,
  factr_tol = 100,
  lmm = 6,
  seed = 1234,
  eval_lp_draws = TRUE
)
}
\arguments{
\item{model}{Stan model (compiled using rstan::stan_model())}

\item{data}{data list for call to rstan::sampling for specified model}

\item{init_bound}{the boundwith of random initials for each dimension (default = 2.0)}

\item{N1}{maxium number of iterations for L-BFGS (default = 1000)}

\item{N_sam_DIV}{number of samples used in estimating ELBO (default = 5)}

\item{N_sam}{number of samples from the approximating Normal returned
will not do extra samples when < N_sam_DIV  (default = 5)}

\item{factr_tol}{the option factr in optim() (default = 1e2)}

\item{lmm}{the option lmm in optim() (default = 6)}

\item{seed}{random seed for generating initials}

\item{eval_lp_draws}{logical; if TRUE (default) evaluate the log-densities of approximating draws}
}
\value{
a list with named elements:
\code{sample_pkg_save}: matrics for estiamted inv Hessian
\code{DIV_save}: approximating draws and DIV
\code{y}: optimization path
\code{fn_call}: no. calls of log density
\code{gr_call}: no. calls of gradient
status
}
\description{
Run one-path Pathfinder for specified Stan model and data for
the specified number of iterations using the specified bound on
uniform initialization on the unconstrained scale.  See opt_path()
for a description of output format and algorithm.
}
