% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{opt_path_stan_parallel}
\alias{opt_path_stan_parallel}
\title{Run one-path Pathfinder with random initials in parallel}
\usage{
opt_path_stan_parallel(
  seed_init,
  seed_list,
  mc.cores,
  model,
  data,
  init_bound = 2,
  N1 = 1000,
  N_sam_DIV = 5,
  N_sam = 100,
  factr_tol = 100,
  lmm = 6,
  eval_lp_draws = TRUE
)
}
\arguments{
\item{seed_init}{array of random seeds for generating random initials}

\item{seed_list}{array of random seeds for running one-path Pathfinders}

\item{mc.cores}{The maximum number of one-path Pathfinder to run in parallel}

\item{model}{Stan model (compiled using rstan::stan_model())}

\item{data}{data list for call to rstan::sampling for specified model}

\item{init_bound}{the boundwith of random initials for each dimension (default = 2.0)}

\item{N1}{maxium number of iterations for L-BFGS (default = 1000)}

\item{N_sam_DIV}{number of samples used in estimating ELBO (default = 5)}

\item{N_sam}{number of samples from the approximating Normal returned
will not do extra samples when < N_sam_DIV  (default = 5)}

\item{factr_tol}{the option factr in optim() (default = 1e2)}

\item{lmm}{the option lmm in optim() (default = 6)}

\item{eval_lp_draws}{logical; if TRUE (default) evaluate the log-densities of approximating draws}
}
\value{
A list of outputs from \code{\link{opt_path}}
}
\description{
Run one-path Pathfinder with random initials in parallel
}
