% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{params_only}
\alias{params_only}
\title{Return optimization path with last column (objective function value)
removed.}
\usage{
params_only(path)
}
\arguments{
\item{path}{optimization path with last column for objective
function value}
}
\value{
Optimization path with last column removed.
}
\description{
Return optimization path with last column (objective function value)
removed.
}
