% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{random_sample_Each}
\alias{random_sample_Each}
\title{random sample for one-path Pathfinder}
\usage{
random_sample_Each(param_path, seed)
}
\arguments{
\item{param_path}{output of function opt_path_stan_parallel()}

\item{seed}{random seed}
}
\description{
random sample for one-path Pathfinder
}
