% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{Sam_N_apx}
\alias{Sam_N_apx}
\title{Generate N_sam samples from the approximating Normal given the output of
subroutine Form_N_apx}
\usage{
Sam_N_apx(sample_pkg, N_sam)
}
\arguments{
\item{sample_pkg}{The output of subroutine Form_N_apx}

\item{N_sam}{Number of samples}
}
\value{
\item{samples}{The samples from the approximating Normal,
each column stores a sample}
\item{lp_approx_draws}{The log-density of generated samples under
approximating Normal}
}
\description{
Generate N_sam samples from the approximating Normal given the output of
subroutine Form_N_apx
}
