% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCStan-class.R
\name{featureNames,BASiCStan-method}
\alias{featureNames,BASiCStan-method}
\alias{sampleNames,BASiCStan-method}
\alias{featureNames<-,BASiCStan-method}
\alias{sampleNames<-,BASiCStan,ANY-method}
\title{Feature and sample names for \linkS4class{BASiCStan} objects.}
\usage{
\S4method{featureNames}{BASiCStan}(object)

\S4method{sampleNames}{BASiCStan}(object)

\S4method{featureNames}{BASiCStan}(object) <- value

\S4method{sampleNames}{BASiCStan,ANY}(object) <- value
}
\arguments{
\item{object}{An object of class \linkS4class{BASiCStan}.}

\item{value}{The value for feature or sample names to be set to.}
}
\value{
For the getters, a character vector.
For the setters, an object of class \linkS4class{BASiCStan}.
}
\description{
Getters and setters for names of \linkS4class{BASiCStan} objects.
}
