% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder_utils.R
\name{updateYS}
\alias{updateYS}
\title{update the matrix for storing history of updates along optimization opath}
\usage{
updateYS(Ykt_h, Ykt, lmm)
}
\arguments{
\item{Ykt_h}{history of updates}

\item{Ykt}{latest update}

\item{lmm}{The size of history}
}
\value{
The updated matrix
}
\description{
update the matrix for storing history of updates along optimization opath
}
