% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCStan-package.R
\docType{package}
\name{BASiCStan-package}
\alias{BASiCStan-package}
\title{The 'BASiCStan' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan.
R package version 2.21.2. https://mc-stan.org
}
