% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCStan.R
\name{BASiCStan}
\alias{BASiCStan}
\title{Stan implementation of BASiCS (with spike-in genes only)}
\usage{
BASiCStan(
  Data,
  Method = c("vb", "sampling", "optimizing"),
  Regression = TRUE,
  WithSpikes = length(altExpNames(Data)) > 0,
  BatchInfo = Data$BatchInfo,
  L = 12,
  ReturnFit = TRUE,
  ...
)
}
\arguments{
\item{Data}{SingleCellExperiment object}

\item{Method}{Inference method: Variational Bayes ("vb") or Hamiltonian
Monte Carlo ("sampling").}

\item{Regression}{Logical scalar indicating whether or perform the
mean-overdispersion Regression described in Eling et al.}

\item{WithSpikes}{Do the data contain spike-in genes? See BASiCS for details.}

\item{...}{Passed to vb or sampling.}
}
\description{
Stan implementation of BASiCS (with spike-in genes only)
}
