% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCStan.R
\name{BASiCStan}
\alias{BASiCStan}
\title{Stan implementation of BASiCS (with spike-in genes only)}
\usage{
BASiCStan(
  Data,
  Method = c("vb", "sampling", "optimizing"),
  WithSpikes = length(altExpNames(Data)) > 0,
  BatchInfo = Data$BatchInfo,
  L = 12,
  PriorMu = c("EmpiricalBayes", "uninformative"),
  ReturnBASiCS = TRUE,
  ...
)
}
\arguments{
\item{Data}{SingleCellExperiment object}

\item{Method}{Inference method: Variational Bayes ("vb") or Hamiltonian
Monte Carlo ("sampling").}

\item{WithSpikes}{Do the data contain spike-in genes? See BASiCS for details.}

\item{BatchInfo}{Vector describing which batch each cell is from.}

\item{L}{Number of regression terms (including slope and intercept) to use in
joint prior for mu and delta.}

\item{PriorMu}{Type of prior to use for mean expression. Default is
"EmpiricalBayes", but "uninformative" is the prior used in Eling et al. and
previous work.}

\item{ReturnBASiCS}{Should the object be converted into a
\linkS4class{BASiCS_Chain} object?}

\item{...}{Passed to vb or sampling.}
}
\description{
Stan implementation of BASiCS (with spike-in genes only)
}
\examples{
library("BASiCS")
sce <- BASiCS_MockSCE()
BASiCStan(sce, tol_rel_obj = 1e-1)
}
